modules/plugin/test/mochitest/test_crash_nested_loop.html

<head>
  <title>Plugin crashing in nested loop</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

<body>
  <iframe id="iframe1" src="crashing_subpage.html" width="600" height="600"></iframe>

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var iframe = document.getElementById('iframe1');

  window.frameLoaded = function frameLoaded_toCrash() {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    var prefs = Components.classes['@mozilla.org/preferences-service;1']
      .getService(Components.interfaces.nsIPrefBranch);
    if (!prefs.getBoolPref('dom.ipc.plugins.enabled')) {
      ok(true, "Skipping this test when IPC plugins are not enabled.");
      SimpleTest.finish();
      return;
    }

    var p = iframe.contentDocument.getElementById('plugin1');

    // This test is for bug 550026, which is inherently nondeterministic.
    // If we hit that bug, the browser process would crash when the plugin
    // crashes in crashInNestedLoop.  If not, we'll pass "spuriously".
    try {
      p.crashInNestedLoop();
      // The plugin didn't crash when expected.  This happens sometimes.  Give
      // it longer to crash.  If it crashes (but not at the apropriate time),
      // soft fail with a todo; if it doesn't crash then something went wrong,
      // so fail.
      setTimeout(
        function() {
          try {
            p.getPaintCount();
            ok(false, "plugin should crash");
          } catch (e) {
            todo(false, "p.crashInNestedLoop() should throw an exception");
          }

          SimpleTest.finish();
        },
        1000);
    }
    catch (e) {
      ok(true, "p.crashInNestedLoop() should throw an exception");
      SimpleTest.finish();
    }

  }

  </script>
